+++ api/admin/internal/logic/member/level/memberlevelupdatelogic.go
@@ -32,7 +32,7 @@ func (l *MemberLevelUpdateLogic) MemberLevelUpdate(req types.UpdateMemberLevelRe
 		LevelName:          req.Name,
 		GrowthPoint:        req.GrowthPoint,
 		DefaultStatus:      req.DefaultStatus,
-		FreeFreightPoint:   float32(req.FreeFreightPoint),
+		FreeFreightPoint:   req.FreeFreightPoint,
 		CommentGrowthPoint: req.CommentGrowthPoint,
 		IsFreeFreight:      req.IsFreeFreight,
 		IsSignIn:           req.IsSignIn,
