+++ api/admin/internal/logic/order/cart/cartitemaddlogic.go
@@ -31,7 +31,7 @@ func (l *CartItemAddLogic) CartItemAdd(req types.AddCartItemReq) (*types.AddCart
 		ProductSkuId:      req.ProductSkuId,
 		MemberId:          req.MemberId,
 		Quantity:          req.Quantity,
-		Price:             float32(req.Price),
+		Price:             req.Price,
 		ProductPic:        req.ProductPic,
 		ProductName:       req.ProductName,
 		ProductSubTitle:   req.ProductSubTitle,
