+++ rpc/pms/internal/logic/productservice/productaddlogic.go
@@ -107,7 +107,7 @@ func (l *ProductAddLogic) ProductAdd(in *pmsclient.ProductAddReq) (*pmsclient.Pr
 		_ = memberPrice.Create(&model.PmsMemberPrice{
 			ProductID:       productId,
 			MemberLevelID:   list.MemberLevelId,
-			MemberPrice:     float64(list.MemberPrice),
+			MemberPrice:     list.MemberPrice,
 			MemberLevelName: list.MemberLevelName,
 		})
 	}
@@ -118,8 +118,8 @@ func (l *ProductAddLogic) ProductAdd(in *pmsclient.ProductAddReq) (*pmsclient.Pr
 		_ = ladder.Create(&model.PmsProductLadder{
 			ProductID: productId,
 			Count:     list.Count,
-			Discount:  float64(list.Discount),
-			Price:     float64(list.Price),
+			Discount:  list.Discount,
+			Price:     list.Price,
 		})
 	}
 	//4.满减价格
@@ -127,8 +127,8 @@ func (l *ProductAddLogic) ProductAdd(in *pmsclient.ProductAddReq) (*pmsclient.Pr
 	for _, list := range in.ProductFullReductionList {
 		_ = full.Create(&model.PmsProductFullReduction{
 			ProductID:   productId,
-			FullPrice:   float64(list.FullPrice),
-			ReducePrice: float64(list.ReducePrice),
+			FullPrice:   list.FullPrice,
+			ReducePrice: list.ReducePrice,
 		})
 	}
 
@@ -141,12 +141,12 @@ func (l *ProductAddLogic) ProductAdd(in *pmsclient.ProductAddReq) (*pmsclient.Pr
 		_ = sku.Create(&model.PmsSkuStock{
 			ProductID:      productId,
 			SkuCode:        list.SkuCode,
-			Price:          float64(list.Price),
+			Price:          list.Price,
 			Stock:          list.Stock,
 			LowStock:       list.LowStock,
 			Pic:            list.Pic,
 			Sale:           list.Sale,
-			PromotionPrice: float64(list.PromotionPrice),
+			PromotionPrice: list.PromotionPrice,
 			LockStock:      list.LockStock,
 			SpData:         list.SpData,
 		})
