+++ script/sql/oms/oms_order_operate_history.sql
@@ -5,10 +5,11 @@ create table oms_order_operate_history
     order_id     bigint       not null comment '订单id',
     operate_man  varchar(100) not null comment '操作人：用户；系统；后台管理员',
     create_time  datetime     not null comment '操作时间',
-    order_status int(1)       not null comment '订单状态：0->待付款；1->待发货；2->已发货；3->已完成；4->已关闭；5->无效订单',
+    order_status tinyint not null comment '订单状态：0->待付款；1->待发货；2->已发货；3->已完成；4->已关闭；5->无效订单',
     note         varchar(500) null comment '备注'
 )
     comment '订单操作历史记录';
+
 
 INSERT INTO oms_order_operate_history (id, order_id, operate_man, create_time, order_status, note) VALUES (5, 12, '后台管理员', '2018-10-12 14:01:29', 2, '完成发货');
 INSERT INTO oms_order_operate_history (id, order_id, operate_man, create_time, order_status, note) VALUES (6, 13, '后台管理员', '2018-10-12 14:01:29', 2, '完成发货');
