+++ script/sql/pms/pms_comment_replay.sql
@@ -7,7 +7,7 @@ create table pms_comment_replay
     member_icon      varchar(255)  not null,
     content          varchar(1000) not null,
     create_time      datetime      not null,
-    type             int(1)        not null comment '评论人员类型；0->会员；1->管理员'
+    type tinyint not null comment '评论人员类型；0->会员；1->管理员'
 )
     comment '产品评价回复表';
