+++ script/sql/pms/pms_product_attribute.sql
@@ -4,17 +4,18 @@ create table pms_product_attribute
         primary key,
     product_attribute_category_id bigint       not null,
     name                          varchar(64)  not null,
-    select_type                   int(1)       not null comment '属性选择类型：0->唯一；1->单选；2->多选',
-    input_type                    int(1)       not null comment '属性录入方式：0->手工录入；1->从列表中选取',
+    select_type     tinyint not null comment '属性选择类型：0->唯一；1->单选；2->多选',
+    input_type      tinyint not null comment '属性录入方式：0->手工录入；1->从列表中选取',
     input_list                    varchar(255) not null comment '可选值列表，以逗号隔开',
     sort                          int          not null comment '排序字段：最高的可以单独上传图片',
-    filter_type                   int(1)       not null comment '分类筛选样式：1->普通；1->颜色',
-    search_type                   int(1)       not null comment '检索类型；0->不需要进行检索；1->关键字检索；2->范围检索',
+    filter_type     tinyint not null comment '分类筛选样式：1->普通；1->颜色',
+    search_type     tinyint not null comment '检索类型；0->不需要进行检索；1->关键字检索；2->范围检索',
     related_status  tinyint not null comment '相同属性产品是否关联；0->不关联；1->关联',
     hand_add_status tinyint not null comment '是否支持手动新增；0->不支持；1->支持',
     type            tinyint not null comment '属性的类型；0->规格；1->参数'
 )
     comment '商品属性参数表';
+
 
 INSERT INTO pms_product_attribute (id, product_attribute_category_id, name, select_type, input_type, input_list, sort, filter_type, search_type, related_status, hand_add_status, type) VALUES (1, 1, '尺寸', 2, 1, 'M,X,XL,2XL,3XL,4XL', 0, 0, 0, 0, 0, 0);
 INSERT INTO pms_product_attribute (id, product_attribute_category_id, name, select_type, input_type, input_list, sort, filter_type, search_type, related_status, hand_add_status, type) VALUES (7, 1, '颜色', 2, 1, '黑色,红色,白色,粉色', 100, 0, 0, 0, 1, 0);
